requirements.txt
Django==2.1.2
django-bootstrap==0.2.4
django-bootstrap-form==3.4
djangorestframework==3.9.0
gunicorn==19.9.0
mysqlclient==1.3.12
Pillow==5.3.0
psycopg2==2.7.6.1
pytz==2018.5
psycopg2-binary==2.7.6.1
django-rest-framework==0.1.0
reportlab==3.5.11